requirements.txt
streamlit
kafka-python
psycopg2-binary
google-generativeai
pandas
sqlalchemy